requirements.txt
streamlit
torch
torchvision
pandas
numpy
scikit-learn
matplotlib
seaborn
Pillow